test sparse ops, extract to separate files
